requirements.txt
yfinance
matplotlib
schedule
python-dotenv
discord.py